rename: 카멜 케이스 -> 스네이크 케이스 (#11)
