+++ src/api/user_api.js
@@ -1,5 +1,4 @@
 import axios from 'axios';
-
 const BACKEND_URL = import.meta.env.VITE_BACKEND_URL;
 
 export const verifyUserBySchoolNumber = async (schoolnumber) => {
