+++ src/components/input_field.css
@@ -1,0 +1,30 @@
+.input-field {
+	position: absolute;
+	width: 220px;
+	height: 45px;
+	left: 50%;
+	transform: translateX(-50%);
+	background: #fff;
+	border: 2px solid #d9d9d9;
+	border-radius: 10px;
+	padding: 0 10px;
+	font-size: 14px;
+}
+
+.input-field::placeholder {
+	text-align: center;
+}
+
+.input-field:focus {
+	outline: none;
+	border-color: #2e5bff;
+	box-shadow: 0 0 0 2px rgba(46, 91, 255, 0.2);
+}
+
+.input-field.student-id {
+	top: 190px;
+}
+
+.input-field.password {
+	top: 250px;
+}
