+++ src/pages/main/main_page.jsx
@@ -1,4 +1,4 @@
-import MainForm from './main-form';
+import MainForm from './main_form';
 import './main.css';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { useEffect } from 'react';
